test_projects/qwen_qwen_2_5_coder_32b_instruct_python/requirements.txt
Flask==2.0.2
Werkzeug==2.0.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
